move eval script
